Refactoring and Physics

+++ src/main/java/Components/GridLines.java
@@ -18,22 +18,22 @@ public class GridLines extends Component
         Vector2f projectionSize = camera.GetProjectionSize();
 
         // Convert the camera position into a multiple of grid_width and grid_height
-        int firstX = ((int)(cameraPos.x / Settings.GRID_WIDTH) - 1) * Settings.GRID_WIDTH;
-        int firstY = ((int)(cameraPos.y / Settings.GRID_HEIGHT) - 1) * Settings.GRID_HEIGHT;
+        float firstX = ((int)(cameraPos.x / Settings.GRID_WIDTH) - 1) * Settings.GRID_WIDTH;
+        float firstY = ((int)(cameraPos.y / Settings.GRID_HEIGHT) - 1) * Settings.GRID_HEIGHT;
 
         // Getting the total number of horizontal and vertical lines
         int numVertLines = (int)((projectionSize.x * camera.GetZoom()) / Settings.GRID_WIDTH) + 2;
         int numHorLines = (int)((projectionSize.y * camera.GetZoom()) / Settings.GRID_HEIGHT) + 2;
 
-        int width = (int)(projectionSize.x * camera.GetZoom()) + Settings.GRID_WIDTH * 2;
-        int height = (int)(projectionSize.y * camera.GetZoom()) + Settings.GRID_HEIGHT * 2;
+        float width = (int)(projectionSize.x * camera.GetZoom()) + Settings.GRID_WIDTH * 2;
+        float height = (int)(projectionSize.y * camera.GetZoom()) + Settings.GRID_HEIGHT * 2;
 
         int maxLines = Math.max(numHorLines, numVertLines);
         Vector3f color = new Vector3f(0.1f, 0.1f, 0.1f);
         for (int i = 0; i < maxLines; i++)
         {
-            int x = firstX + (Settings.GRID_WIDTH * i);
-            int y = firstY + (Settings.GRID_HEIGHT * i);
+            float x = firstX + (Settings.GRID_WIDTH * i);
+            float y = firstY + (Settings.GRID_HEIGHT * i);
 
             if (i < numVertLines)
                 DebugDraw.AddLine2D(new Vector2f(x, firstY), new Vector2f(x, firstY + height), color);
